requirements.txt
# Setup tools (install these first for Python 3.12)
setuptools==69.0.3
wheel==0.42.0

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0

# Video processing
opencv-python==4.8.1.78
pillow==10.1.0
imageio==2.33.0
imageio-ffmpeg==0.4.9

# AI/ML dependencies
openai==1.6.1
# Temporarily commenting out langchain due to compatibility issues
# langchain==0.1.0
# langchain-openai==0.0.2
# chromadb==0.4.20
tiktoken==0.5.2

# Web and async
aiofiles==23.2.1
websockets==12.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1
